backend/requirements.txt
fastapi
uvicorn
appwrite==15.3.0
pydantic[email]
pydantic-settings
bcrypt
PyJWT
firebase-admin
